+++ cmd/hidden_lake/applications/messenger/pkg/app/app_test.go
@@ -44,7 +44,6 @@ func TestApp(t *testing.T) {
 	cfg, err := config.BuildConfig(tcPathConfig, &config.SConfig{
 		FSettings: &config.SConfigSettings{
 			FMessagesCapacity: testutils.TCCapacity,
-			FWorkSizeBits:     testutils.TCWorkSize,
 		},
 		FAddress: &config.SAddress{
 			FInterface: testutils.TgAddrs[57],